requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic<2
PyJWT
cryptography
pymysql
redis
python-dotenv
email-validator
APScheduler
requests
razorpay
openpyxl==3.1.5
alembic
boto3
python-multipart
bcrypt
# Google API packages for gmeet_api
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
# "pydantic[email]"

#pip install -r requirements.txt

# cd nucleotide_backend

# python -m venv venv      [environment creation]

# venv\Scripts\activate    [windows]
# source venv/bin/activate [macOS]

# uvicorn app.main:app --reload [to run the app]